move slice functions to new file
